+++ src/main/java/chaminaTech/Repository/SuprimentoRepository.java
@@ -5,7 +5,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
+import java.math.BigDecimal;
+
 public interface SuprimentoRepository extends JpaRepository<Suprimento, Long> {
     @Query("SELECT COALESCE(SUM(s.valor), 0) FROM Suprimento s WHERE s.caixa.id = :caixaId")
-    Double findTotalSuprimentosByCaixaId(@Param("caixaId") Long caixaId);
+    BigDecimal findTotalSuprimentosByCaixaId(@Param("caixaId") Long caixaId);
 }
